ext-session-lock-v1: relicense to MIT
All other protocols in wayland-protocols are released under the MIT
license and this one was only merged with the ISC license by accident.

I am the only person who has touched this protocol in commits and the
only copyright holder, so relicensing to bring this protocol in line
with the rest is easy in this case.

References: https://gitlab.freedesktop.org/wayland/wayland-protocols/-/issues/119
Signed-off-by: Isaac Freund <mail@isaacfreund.com>