xdg-output: Remove and tweak contradicting examples
The "logical size" as stated by the first paragraph corresponds to the
monitor size in the global compositor space.

To most clients, this is unnecessary information, and should be ignored,
but some used the listed examples to derive information that contradicts
the very definition of what this event communicates.

One example tried to add surface size assumptions, which was not
correct. Remove this part completely, clients should not try to
configure their surface sizes from the logical size of a monitor.

The other is the list of size examples; it tried to communicate that a
compositor sometimes may not scale the viewport of the monitor in its
global compositor coordinate space, in which case, the logical size
itself matches the actual resolution. Tweak this wording to make that
clear that it does not related to any surface size.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>