xdg-shell: Add suspended toplevel state
Add a toplevel state to indicate that surface repaints have been
suspended. This may arise due to occlusion, output power state, etc.

In this state, clients can choose to take meaningful action such as
suspending any processing which would drive a repaint loop, or
communicating to the active browser tab that the tab is not
system-visible, or any other action that would be taken by a client not
expecting to repaint until further notice.

cf. discussion in wayland/wayland-protocols!99

Signed-off-by: Daniel Stone <daniels@collabora.com>