stable/xdg-shell: clarify initial wl_surface acknowledgement
Clarify how and when initial wl_surface state provided by the core
protocol or by extensions to the wl_surface, like as
wp_fractional_scale_v1, is being delivered.

The motivation for such change is to make it clear that the first frame
for xdg-shell will be perfect, which implies that scaling and similar
properties affecting presentation would be delivered in time.

Signed-off-by: Kirill Chibisov <contact@kchibisov.com>