Add staging system bell protocol
This is meant to let applications ring the system bell. It needs to be a
Wayland protocol because a system bell is not necessarily audiable; for
for example accessibility reasons, it might need be a visual feedback,
which may be tied to a specific window. Accessibility features are
usually configured globally, and one likely wants identical visual
feedback for all system bell ringings, so it doesn't fit as a client
side only feature.

This aims to replaced and deprecate the `gtk_shell1.system_bell`
request.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>