staging: Add xdg-toplevel-icon protocol for dedicated toplevel icons
This protocol allows clients to set icons for their toplevel windows.
Icons can be loaded from the XDG icon stock using their name, or can
alternatively be provided by the client as wl_shm-backed wl_buffer.

A toplevel icon represents the individual toplevel (unlike the
application or launcher icon, which represents the application as a
whole), and may be shown in window switchers, window overviews and
taskbars that list individual windows.

Resolves: #52

Signed-off-by: Matthias Klumpp <matthias@tenstral.net>