xdg-toplevel-icon: clarify that wl_buffer.release is unused
The protocol as-is doesn't allow clients to mutate wl_buffers.
Let's make it clear that wl_buffer.release is not used for that
purpose. Buffer re-use can be added in a future protocol version
if desirable.

Add a small note to explain that no wl_buffer mutation implies no
wl_shm_pool's backing storage mutation as well.

Signed-off-by: Simon Ser <contact@emersion.fr>
Closes: https://gitlab.freedesktop.org/wayland/wayland-protocols/-/issues/201