xdg-shell: clarify clients set their initial state before initial commit
It wasn't clear that it's important clients set up their initial
xdg_surface state before they send the initial commit. This is
required for the compositor to be able to send a proper configure
event depending on size constraints and any policies it might want
to apply (e.g. specific app ID always shows up in a designated
workspace).

Signed-off-by: Simon Ser <contact@emersion.fr>